requirements.txt
Django==2.2.18
PyMySQL==0.9.3
coreapi==2.3.3
coreschema==0.0.4
django-filter==2.3.0
django-redis==4.12.1
django-celery==3.3.1
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
djangorestframework-bulk==0.2.1
Markdown==3.3.4
django-rest-swagger==2.2.0
drf-yasg==1.17.1
django-cors-headers==3.4.0
uritemplate==3.0.1
PyYAML==5.4
itsdangerous==1.1.0
celery==4.4.7
pytz==2018.9
redis==3.5.3
oss2=2.17.0
django-mptt==0.14.0
channels==3.0.4
channels-redis==3.2.0
django-elasticsearch-dsl==7.3
elasticsearch-dsl==7.4.1